renaming to metacat-canonical-converters module
