+++ metacat-canonical-converters/src/main/java/com/netflix/metacat/canonical/type/package-info.java
