+++ metacat-canonical-converters/src/main/java/com/netflix/metacat/canonical/typeconverters/TypeUtil.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2016 Netflix, Inc.
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package com.netflix.metacat.canonical.typeconverters;
+
+import com.google.common.base.Preconditions;
+
+/**
+ * Created by zhenli on 12/19/16.
+ */
+public final class TypeUtil {
+    private TypeUtil() {
+    }
+
+    /**
+     * CheckType.
+     * @param value value
+     * @param target type
+     * @param name name
+     * @param <A> A
+     * @param <B> B
+     * @return B
+     */
+    public static <A, B extends A> B checkType(final A value, final Class<B> target, final String name) {
+        Preconditions.checkNotNull(value, "%s is null", name);
+        Preconditions.checkArgument(target.isInstance(value),
+            "%s must be of type %s, not %s",
+            name,
+            target.getName(),
+            value.getClass().getName());
+        return target.cast(value);
+    }
+}
